+++ database/migrations/2016_05_10_163028_create_table_biaya_satwa.php
@@ -18,7 +18,7 @@ class CreateTableBiayaSatwa extends Migration
             $table->integer('id_responden');
 
             $table->integer('volume')->nullable();
-            $table->float('harga_beli')->nullable();
+            $table->double('harga_beli')->nullable();
             $table->integer('umur_ekonomis')->nullable();
             
             $table->softDeletes();
