+++ database/migrations/2016_05_11_101303_create_table_biaya_wisata.php
@@ -17,10 +17,10 @@ class CreateTableBiayaWisata extends Migration
             $table->integer('id_responden');
 
             $table->integer('jenis_pengeluaran')->nullable();
-            $table->float('biaya')->nullable();
-            $table->float('jumlah')->nullable();
+            $table->double('biaya')->nullable();
+            $table->double('jumlah')->nullable();
             $table->string('satuan_jumlah', 30)->nullable();
-            $table->float('total')->nullable();
+            $table->double('total')->nullable();
             
             $table->softDeletes();
             $table->timestamps();
